Implement architecture for JavaScript-calling-Java thread-safety
JavaScript accesses Java on a separate thread associated with the
WebView that is different than the main UI thread.

Approach is to use a Handler object when processing Java function
from JavaScript, and have handler post to message queue (of
a worker thread) so the messages or runnables execute sequentially.

Minor refactoring to inline the TextToSpeech.OnInitListener() with
an anonymous inner class.