Add textview and buttons to activity layout Refactor processing of notes json object from ajax call
We don't want the app to be just a WebView, so add TextView
and couple of Buttons to the layout; they get, display,
and clear display, the first note entry from the web app.

Refactor processing of the json notation String returned by
evaluateJavascript() from the JavaScript ajax call. The
array json notation obtained in the callback is converted
into a Java JSONArray object for easier processing.