Add server-side code for the web app in WebView
Add contents of 'server' directory for MyMobileOrganizer taken from
Pro iOS and Android Apps for Business.

In DAO.js:
 Fix schema for 'note' that has extra field that does not appear to
be used since it is not on the input form.
 Fix minor typo in console log message.

Keep the original server IP address and port number because it seems
possible to use Chrome port forwarding as described here
 https://developer.chrome.com/devtools/docs/remote-debugging#enable-reverse-port-forwarding
In dev machine, chrome://inspect page, use setting of: Device: 8080, and Host: 127.0.0.1:80
and set code in Android for ajax url to: http://127.0.0.1:8080.