Remove karura project as library, add the needed class files
Since we don't use the karura framework per se, it seems overkill
to make the entire framework a dependent library for this project
 https://github.com/karuradev/karura

We actually only use two class files from the karura framework; it
is just convenient to piggy-back on those two classes to access the
Contacts data on the device, which is tied to the built-in People
native app. So we add only those two class files (plus a couple more
which seem to be dependencies), and remove karura project as a
library project.