Replace our own MyWorkerThread class by using HandlerThread class
Android's HandlerThread already has a Looper, so can avoid hooking up one.
(Leaving MyWorkerThread class as example of adding a Looper to a Thread.)

Improve comments for the two Java methods that are bound to WebView
JavaScript and so run in its separate thread for WebView's JavaScript/Java
binding.

Add some more code to try to prevent leaks.