Target sdk version 21 Lollipop
Update manifest to change targetSdkVersion to 21 from 19,
for Nexus 4 that just got OTA upgrade to Lollipop.

Remove the MainActivity label for activity.
For some reason, the label for the MAIN and LAUNCHER intent-filter
is no longer overriding the activity label and the app icon label
in launcher is MainActivity instead of the desired app name.
Also remove the label for intent-filter since it is now not needed.